Update for KSP 1.7.9
Work fine on KSP 1.7.3, without any visible error or crash and produce the desired effects. So I propose to update it to the latest KSP version to enable CKAN user to use it without the need for "KSP version compatibility override".

I also update LineRender initialization to remove the deprecated function calls.